[AS7-5884] Maintain a reference count for deployment hashes deployed to the content repository.
This is to handle the case where the same deployment is deployed with the same name, to make sure
that a copy is kept in the repository if one is removed. Once all deployments are removed, the
content is removed.

was: 66c57881c9b55c0f2d467cbf518b13dc91b3975d