AS7-5971 Work around transient dependencies problem by waiting for all additional modules to be availble before loading any deployment modules.
This is not a 100% fix, but as the problem is a bit of an edge case as it is,
I think this should be sufficient. Unfortunalty there is no correct way to fully express
module dependencies using services, due to curcular references and re-exported dependencies.

was: dbd11d9bf56e3d1e4e142a8b7349054afc23a611