[AS7-6873] NPE when calling a runtime operation on non-existing resource
before executing the runtime step, call
context.readResource(EMPTY_ADDRESS, false) to ensure that there is a
resource. If that's not the case, the operation will fail with a
descriptive message instead of reporting an exception later on (such as
a NPE)

In most cases, a runtime operation will be executed for a resource that
exists. In the few case where there is no resource (such as a messaging
core-queue created by HornetQ core protocol), the resourceMustExist()
method can be overridden to bypass the check on the resource existence

was: da1f5d2b70f8b890a9ce6bee069eef076e5aad00