[WFLY-2099] / [WFLY-490] Add a duplicate check with appropriate error message for duplicate include / exclude definitions.
As the address of the include/exclude is not suitable to check for duplicates we need to also check the existing model.

Note: At the time the model is checked if there is a duplicate we will find the entry twice.

was: f3c7110f6949d3499fb33f9814e7acf4243e0009