[WFLY-2730] HTTP upgrade multiplexing
When the HTTP upgrade handshake is performed, check whether
the http-upgrade-endpoint is defined and accepts the handshake
only if its value matches the acceptorName.
This http-upgrade-endpoint header is used by a http-connector to
specify which http-acceptor must be used in case there are multiple
http-acceptors registered on the server.

If the http-upgrade-endpoint header is not present, the 1st http-acceptor
encountered will be used. This is not deterministic and is suitable only
if there is a single http-acceptor.

* remove native messaging port for messaging subsystem.xml configuration
and use http-connector for JMS RemoteConnectionFactory
* add a http-connector-throughput to replace the corresponding
  netty-throughput

JIRA: https://issues.jboss.org/browse/WFLY-2730

was: 7bfe28abd5b7dd3b3c40a5d7531b66357b9f4d99