Remove PersistentResourceDefinitionBuilder.addAttributes(Collection)
This version of the method is problematic because any changes to the underling
list will cause existing parsers to change. In particular if an attribute is added
or removed this change will affect old versions of the parser.

was: cb7ae82f9f0a69586fccf4c98df693c2653c911f