[WFCORE-208] : Expose 'singleton' type information in read-children-types
If the "include-singletons" parameter is set to "true", the operation result will include the full key/value pair for any singleton registration.
If "include-singletons" is "true" a non-fully-qualified type (i.e. just key instead of key=value) will only appear in the result if there is a wildcard registration for that key.