[WFCORE-452] Rework ModelControllerServiceInitialization to boot op.
Previously the ModelControllerServiceInitialization loading was
hardcoded as part of the boot sequence. However, in order to use JMX as
a subsystem on the HC we need access to the services started by JMX's
MCSI. Putting these earlier also does not work, since patching's MCSI
needs to read values from the model, which do not get set until the boot
time. Hence this initialization has been refactored to an extra boot
RUNTIME step which has access to the model, and is able to install the
services JMX needs on a host.