[WFCORE-297] : HC should remember the 'run' state of the server instances after crash or shutdown
Since starting or stopping a server is a read-only operation this can't be done as proposed.
When the update-auto-start-with-server-status attribute is set to true on the server declaration then when the server is started a file servername.started will be created in ${jboss.datahome}/auto-start.
When the server is stopped the servername.started will be deleted and a servername.stopped will be created.
Those files are used at startup to determine the auto-start value : if the  update-auto-start-with-server-status is set then the exisitng file will be used instead of the value of the auto-start attribute in the XML.