[WFCORE-401] Test and fix deployment operation diffs
Needed to tweak the order of the operation comparison since the domain
root element children have dependencies between themselves.
For read-only attributes that appear in an add (e.g. deployment's
content), if there is a change drop and re-add the resource.
Execute all the non-extension operations in a composite, with the
removes happening before the adds to support the drop+readd case.
Use an operation header to be able to turn off validation for the
remove (if it comes from a drop+readd the resources will be readded
anyway).
Implement the logic to pull down content from the master if the
deployment contents changed.