[WFCORE-833] Capability-based model reference validation for profiles, socket-binding-groups, server-groups and server-configs
Involves changes to AbstractControllerService/ModelControllerImpl/OperationContextImpl to allow partial deferral of capability resolution to account for the fact that during HostController boot the entire set of configured capabilities is not known until the last of several operation executions is complete (i.e. domain.xml content is not known when host.xml content is installed.)

Also includes a fair bit of test fixture tweaks related to supporting capability resolution testing in core-model-test.