[WFCORE-891] jboss-deployment-structure.xml resource root allows references outside the deployment
I tried to cover these two use cases:
1) If the project is deployed exploded, then you can reference
outside the deployment; check whether the reference is a child
of the deployment root

2) the project is an archive. In this case you cannot reference outside
the deployment (VFS Zip file system) but you can reference a
non-existant resource-root; check if the file exist or not.

Note: when the JAR is inside the deployment, this is processed by
EarStructureProcessor. So if it is a jar/sar is mounted in that stage.