[WFCORE-903] : Resubmit fixed version of WFCORE-860.
Fix consist in release the lock when the notification is processed.
WFCORE-860: Use notifications to help the scanner get a proper status of all deployments between scans
When a deployment is deployed or undeployed from out of a scanner, a notification is sent to the scanner so it can update the status.

Conflicts:
	server/src/main/java/org/jboss/as/server/controller/resources/DeploymentAttributes.java