[WFCORE-949] : Removing http-interface autocompletes but fails.
Removing the management interfaces will put the server in reload mode instead of stopping (and removing) those interfaces at runtime.
Until the introduction a new capability by refactoring the management we check manually if JMX requires a management interface for remoting access and if one is still available after removing.
Fixing the sybsystem test case capability registration so that we may define a serviceValueType.