[WFCORE-992] fix memory leak in host controller
This commit makes sure that the Future objects added to Sets inside
the ManagedServerProxy.activeRequests Map are also properly removed.
The problem was that a delegating Future calls addListener directly
on the underlying Future, so that the callbacks don't get the original
Future, but the underlying one. And this underlying future was
"removed" from the Set (which is a noop, because the underlying Future
was never in the Set in the first place).