[WFCORE-1270] Avoid NPE
GlobalTransformerRegistry.registerTransformer will pass a 'null' into the getOrCreate method that creates a OperationTransformerRegistry, so having this var be 'null' is possible via a normal code path. Why this doesn't regularly fail is a mystery though. Perhaps a different OperationTransformerRegistry instance with the same address/version is created via a different code path in some cases, and there is some timing effect involved. The transformer registration is all single threaded AFAIK though.