[WFCORE-307]: When Wildfly fails with JBAS015810: failed to resolve interface management, it should terminate.
Changing the way a server boots by moving all management and socket related operations to a first batch that will stop the boot process if there is an error.
The other operations subsystems mostly) are in a second batch that won't stop the server from booting.
Fixing some tests according to this new paradigm.
Adding a hack in the ModelControllerImpl for Core Model Tests as we don't have services ready in those tests and they would be breaking with this change.