[WFCORE-2109] Require Maven 3.3.1+ and introduce mvnw
* Maven 3.3.1+ is required by the Enforcer Plugin

* mvnw (a.k.a. Maven Wrapper) is a script that downloads and installs
(if necessary) the required Maven version to ~/.m2/wrapper and runs it
from there. "mvn -N io.takari:maven:wrapper" was used to create all the
necessary files. The same command can be used in the future to upgrade
to a newer wrapper version. Maven version used by the wrapper is stored
in .mvn/wrapper/maven-wrapper.properties

* .mvn/jvm.config is a place where Maven 3.3.1+ looks for JVM parameters
(memory, etc.) which were traditionally defined in the MAVEN_OPTS
environment variable.