[WFCORE-2094] Domain server notifies HC upon controller instability
This commit introduces a notification mechanism via which a listener for "controller instability" notifications can be created and wired into the ModelController.

In a domain server, such a listener is registered. In this case it's an additional function of HostControllerClient, which responds to notifications by asynchronously sending a message back to the managing HC.

Currently the managing HC simply logs a WARN when receiving such a message. In the future it could also do something else with the data, e.g. expose it over the management API for use by the admin console.

Notifications are currently sent in two cases:

1) An OperationContext is interrupted or times out waiting for MSC stability during operation finalization. This means MSC was not stable at the time of finalization and the state of the system is no longer knowable.
2) An Error is thrown during operation execution. Again in such a situation the state of the system is no longer knowable.

In the latter case, there is no guarantee that the listener will be able to send the message to the HC. For example if the Error is an OOME sending a message may not be possible. So this is not a foolproof means of ensuring the HC is aware of server instability. Possible future work to cover that kind of scenario could be to have the HC send some sort of health check request to the server when it detects that an operation has timed out.

This also doesn't send a notification if the server fails to shut down properly, e.g. from problems with the MSC ServiceContainer not stopping. That would be detected by the bootstrap shutdown hook, which has no relationship to this logic.