[WFCORE-2229]: Create serverside operations to redeploy deployments affected by an overlay.
Adding support for a new redeploy-affected method on deployment-overlays.
In domain mode, this is using transformers instead of the classic handlers: thus the operation is transofrmed in a redeploy operation for the target server-groups.
This impacts the ServerResolverOperation to convert locally the transformed operation into the real redeploy of deployments.
Adding a flag to redeploy affected when removing an overlay link.