[WFCORE-2270] Part 1. Make it configurable by the capability designer whether to allow a single registration of a runtime (not 'possible') capability with the same name and scope
In this commit, the default remains 'true', to allow existing cases that need that behavior to still work (see WFLY-8040). However the expectation is those use cases will take advantage of this new API to explicitly configure 'true', and then the default behavior will be switched to 'false' in a later commit.