[WFCORE-3041] Drop the custom policy element parsing we used in WildFly Core 3.0.6 for the 1.1 and 1.0 xsds  and use PersistentResourceXmlDefinition.
This is meant to make it easier to merge the WFCORE-2882 work on top of the WFCORE-3041 work. To this end the policy parsing logic has been moved to a temporary new class 'PolicyParserTemp' and the legacy 'PolicyParser' class which is no longer used, has been restored to its state prior to any WFCORE-3041 changes. (This avoid difficult to untangle merge conflicts with the WFCORE-2882 work.)

Moving to PersistentResourceXmlDefinition loses some custom 'forgiveness' logic we had in 3.0.6 as a compatibility aid. I think this is ok; any need for such forgiveness was always an extreme edge case, and by the time WildFly Core 4 is released any such xml should long since have been converted.