[WFCORE-3415] Faster resource tree size calculation for getMBeanCount()
The biggest thing this does is eliminate RBAC checks as a factor in determining whether a resource appears in the count. Which means, in the case of non-addressable resources the total number of mbeans a user could "see", for example via MBeanServerConnection.queryNames, could be less than the number returned by a simultaneous call to getMBeanCount().

The other big thing this does is avoiding doing two complete counts for the jboss.as and jboss.as.expr JMX domains. There should be no difference in counts between those so if both are installed it just counts one and doubles the result.