Fix WFCORE-3565: Check task queue for more tasks
Updates the runQueuedTask() method to always check whether there are any
tasks in the task queue. This prevents tasks from getting stuck in the
task queue forever if they were placed in there while the container was
at the max request limit.

Additionally, the changes to the runQueuedTask() method in this commit
ensure that forced tasks do not get lost anymore if the container is
paused and has reached the request limit.