WFCORE-3584 Stability monitor should be exposed to DUP's
There are some use cases (e.g. shared java:global bindings) that result in services being installed outside of a deployments ServiceTarget. This means that are not added to the stability monitor, which results in intermittent test failures due to a false stability being achieved before all services have come up.
At present it is not possible to access this stability monitor, so there is no way to fix this issue.