WFCORE-3595 SuspendOnSoftKillTestCase can fail intermittently
The issue is that TestUndertowService is designed for testing suspend/resume rather than an actual graceful shutdown. Because it attempts to perform a dispatch() and queue a task rather than just immediately setting a 503 response it is possible that this request will fail with a 500 response if the XNIO executor is in the process of shutting down.
It would be better if this test checked for a 503 before the skip-graceful request is made.