[WFCORE-4238] Add flags to categorize numerical metrics
Add the COUNTER_METRIC and GAUGE_METRIC flags to AttributeAccess.Flag
enum to further specify the type of metrics when an AttributeDefinition
is registered as a Metric.

Specify the COUNTER_METRIC flag for attributes representing counters (if
the flag is absent, the metric is considered as a gauge).

JIRA: https://issues.jboss.org/browse/WFCORE-4238