[WFCORE-4381] Don't stop the DomainTestSupport before relaunching the HCs. Just stop the HCs.
Stopping the DomainTestSupport also closes down the shared client side Endpoint, and the DomainLifecycleUtil objects use those. Closing it down is what led to the need for unusual start handling after the DomainTestSupport.stop() call.

I also clean out the try with resources handling with the client. The client from DomainLifecycleUtil can now be used and doesn't have to be closed by the test method.