WFCORE-4599 Handle wildfly-config.xml better for authentication
This was previously cached in a static, which does not work for
multiple deployments. This uses an empty authentication context
for the default (if it was not explicitly set), then explicitly
sets per-deployment ClassLoader default to make sure the correct
class loader is in use when the config is loaded.