[WFCORE-5037] Pass in the AttributeDefinition where possible to the JMX TypeConverters
Note that for the main work we still use the descriptor. This fix is
mainly to make it possible to check which kind of Object with complex
values we have.

ObjectTypeAttributeDefinition and ObjectMapAttributeDefinition both
return the same r-r-d output for attributes but are different. The first
one has defined key/value pairs where the keys are all known. e.g. the
following would be an allowed value:
{"my-attr" => {"one" => 100L}}

ObjectMapAttributeDefinition on the other hand is a map of unknown keys
containing those entries, e.g. the following would be an allowed value:
{"my-attr" => {"a" => {"one" => 100L}}}

In both cases the r-r-d output for the attribute would look like:

    "description" => "description",
    "attributes" => {"my-attr" => {
        "type" => OBJECT,
        "description" => "map-of-maps",
        "expressions-allowed" => false,
        "required" => false,
        "nillable" => true,
        "value-type" => {"one" => {
            "type" => LONG,
            "description" => "map-of-maps.one",
            "expressions-allowed" => false,
            "required" => true,
            "nillable" => false
        }},
        "access-type" => "metric",
        "storage" => "runtime"
    }},
    ....

So there is no way for JMX yet to know which of the two cases we have.
Passing in the AttributeDefinition is an attempt at figuring that out.