[WFCORE-5489] As a developer, I want to override management attribute values using environment variables
When validating and setting the value for an non-complex attribute, check if there is an env var
that would override its value before doing any expression resolution or
correction.
Complex attributes (object, property, list) does not allow overriding
their value from an env var.

This feature is enabled by the existence of the
`WILDFLY_OVERRIDING_ENV_VARS` environment variable.

The EnvVarAttributeOverrideModelTestCase test is executed in 2 different
execution to test the behaviour with and without the env var that
activates the feature.

JIRA: https://issues.redhat.com/browse/WFCORE-5489

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>