[WFCORE-5709][WFCORE-5705] Make ElytronExpressionResolver useful in Stage.MODEL
Create a seperate ExpressionResolverExtension interface that core expression resolver plugins like ElytronExpressionResolver implement instead of ExpressionResolver

Add a capability that can be used by ExpressionResolverExtension impls to register themselves with the core expression resolver, allowing their use in Stage.MODEL

Separate ExpressionResolverExtension initialization from resolution so initialization failures can be handled separately and not be treated as indicating the extension
regarded a given expression as relevant to it, but failed to resolve it.

Clarify ExpressionResolverExtension resolution exception handling to distinguish user problems from server problems and not throw OperationFailedException.

Update ReadAttributeHandler.ResolveAttributeHandler to understand the ExpressionResolverExtension exceptions and to treat their presence as indication that
a secure expression was present.

Add manualmode test case analogous to the old CustomVaultInModuleTestCase, but using credential expressions.